Move net handler code out of util
